src/main/java/com/yjs/service/impl/UserServiceimpl.java
package com.yjs.service.impl;


import com.yjs.dataobject.User;
import com.yjs.repository.UserRepository;
import com.yjs.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class UserServiceimpl implements UserService {

    @Autowired
    private UserRepository userRepository;

    @Override
    public User findByUsernameAndPassword(String username, String password) {
        return userRepository.findByUsernameAndPassword(username, password);
    }
}
